fix(gateway): forward full exec defaults to HTTP exec; add changelog entry
